Update stops file architecture.
