+++ app/dashboard/stops/[id]/edit/lib/edit-stop-action.ts
@@ -1,29 +1,9 @@
 'use server';
 
 import {revalidatePath} from "next/cache";
-import { Stop } from '@/app/lib/definitions';
 import {createClient} from '@/utils/supabase/server';
 import {z} from "zod";
 import {redirect} from "next/navigation";
-
-export async function fetchStopById(id: string) {
-
-    const supabase = createClient();
-    
-    try {
-        let queryBuilder = supabase
-        .from("stops")
-        .select("id, created_at, name, description, location")
-        .eq('id', id)
-        .single()
-
-        const { data } = await queryBuilder
-        return data as Stop
-    } catch (error) {
-        console.error('Database Error:', error)
-        throw new Error('Failed to fetch agencies')
-    }
-}
 
 const CreateStopFormSchema = z.object({
     id: z.number(),
