+++ app/dashboard/stops/create/page.tsx
@@ -1,5 +1,5 @@
 import Breadcrumbs from "@/app/ui/lines/breadcrumbs";
-import CreateStopForm from "./create-stop-form";
+import CreateStopForm from "./ui/create-stop-form";
 
 export default async function Page() {
